+++ packages/cli/src/index.ts
@@ -1,7 +1,6 @@
-import { hideBin } from "yargs/helpers";
 import { hi18n } from "./command";
 
-const result = hi18n(hideBin(process.argv));
+const result = hi18n();
 result.catch((e) => {
   console.error(e);
   // eslint-disable-next-line no-process-exit